Add support for custom certificates (#111)
* Add support for custom certificates

The parameters are added to the provider config to support it:
- ca_cert
- client_cert
- client_key

You need to have all three passed for the configuration to be picked up,
it will fail otherise. It also overrides the `tls` parameter
(cfg.TLSConfig), if the latter is set.

Tests are not passing (they were passing upstream either), but running
it locally works fine.

* Handle TLS options inside the provider itself

I mistakenly handled the tls argument inside the mysql driver instead of
doing that in the provider. This commit adds another field,
`tls_config_key`, that will be used to hold the key passed to
`registerTLSConfig()`. It overwrites the `tls` option if set and all the
relevate certificates are set as well. This is inline with the behaviour
defined in the driver https://github.com/gallois/terraform-provider-mysql/blob/4396636ec23a483a9916fd7386cbfc9cc399a365/vendor/github.com/go-sql-driver/mysql/dsn.go#L48-L49

* Change custom tls options to be a map

* Remove unnecessary dependencies

Run `make deps`

* Use nested attribute for custom tls

* Add documentation for custom_tls

* Move variable declaration closer to its usage

* Fix documentation formatting

* Fix typo

* Improve custom tls documentation

---------

Co-authored-by: Felipe Gallois <felipe.gallois@gsr.io>