Fixes TF drift for resource group name and query_limit (#6)
This fixes the following drift:
1. Resource group name would always be in quotes when pulled from the
DB. This PR removes the quotes.
2. Query Limit would always be missing `()` parentheses when pulled from
the DB. This PR standardizes around the query_limit inside the
parentheses. A breaking change is that we expect NO parenthesis to be
passed in for the query_limit from now on. Instead, we will wrap the
query_limit inside parentheses in the actual query rather than rely on
the query_limit string being wrapped in parentheses by the caller.

I also got basic tidb tests working and tested the above changes using
`make testtidb7.5.1`

---------

Co-authored-by: Andrew Chen <andrewchen@plaid.com>
Co-authored-by: Zander Hill <zander@xargs.io>